refactor(pending): extract resolver lifecycle primitives into playback/pending.rs
The race-safety reasoning for pending queue rows (insert -> claim -> Tidal
search -> promote -> gc) lived spread across queue.rs and three resolver
functions in routes.rs. Each resolver inlined its own SELECT/UPDATE
SQL for the same pending_* columns; the "WHERE track_id IS NULL"
promotion guarantee was repeated; the resolving_at lock semantics
were re-implemented per call site.

New playback/pending.rs owns the lifecycle primitives:

  - try_claim   - acquire resolving_at lock (NULL -> NOW iff track_id IS NULL)
  - release     - clear the lock, idempotent
  - read_identity - SELECT artist/title/tidal_id_hint for a given queue id
  - current_pending - SELECT the playback_state-current pending row
  - promote     - atomic UPDATE WHERE track_id IS NULL + external-candidate
                  cleanup, returns true iff this caller won the race

The two resolver orchestrators in routes.rs (resolve_pending_row,
resolve_pending_current_queue_item) and the event-emitting wrapper
(promote_pending_row_emit) now delegate the SQL + race-safety details
to this module. routes.rs no longer touches the queue.pending_* or
queue.resolving_at columns directly.

gc_pending_queue stays in playback/queue.rs because its tests live
there and it shares the queue-table fixture; that's a future cleanup,
not a regression.

routes.rs::promote_pending_row_emit shrinks from ~50 lines to 11.
Both promote_pending_row_emit tests pass; all 5 gc_pending tests pass.
