+++ noor-server/src/playback/mod.rs
@@ -4,6 +4,7 @@ pub mod dj_engine;
 pub mod dj_lookahead;
 pub mod gapless;
 pub mod output;
+pub mod pending;
 pub mod player;
 pub mod queue;
 pub mod runtime;
